Use DI Container to initialize Option classes
"WebhookSubscriptionStore" throws a "AccessDenied" exception, because it was trying to create the "BYOB_TokenMap" directory which still points to the default directory "D:/home/data/byob_graphmap".  The directory is configured in the "SetAppSettings()" method of the GraphOption class. This method is called in the "MicrosoftGraphExtensionConfigProvider" class constructor, which gets not called.
Configuring the option class should not be the responsibility of the services, so this was moved to the DI container.